Move shared context to support/ folder
